# backend/generated-api-client/models/login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend API
 * 간단한 사용자 관리 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 로그인 요청
 */
export interface LoginRequest {
    /**
     * 사용자 이메일
     */
    email: string;
    /**
     * 비밀번호
     */
    password: string;
}