# backend/generated-api-client/models/login-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend API
 * 간단한 사용자 관리 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 로그인 응답
 */
export interface LoginResponse {
    /**
     * JWT 액세스 토큰
     */
    accessToken: string;
    /**
     * 토큰 타입
     */
    tokenType: string;
}