# backend/generated-api-client/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend API
 * 간단한 사용자 관리 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 사용자 정보
 */
export interface UserResponse {
    /**
     * 사용자 ID
     */
    id: number;
    /**
     * 사용자 이름
     */
    name: string;
    /**
     * 사용자 이메일
     */
    email: string;
    /**
     * 역할
     */
    role: string;
}