# backend/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Backend API",
        "description": "\uac04\ub2e8\ud55c \uc0ac\uc6a9\uc790 \uad00\ub9ac API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "tags": [
        {
            "name": "Auth",
            "description": "\uc778\uc99d \uad00\ub828 API"
        },
        {
            "name": "User",
            "description": "\uc0ac\uc6a9\uc790 \uad00\ub828 API"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\ub85c\uadf8\uc778",
                "description": "\uc774\uba54\uc77c/\ube44\ubc00\ubc88\ud638\ub85c \ub85c\uadf8\uc778\ud558\uc5ec JWT \ud1a0\ud070\uc744 \ubc1c\uae09\ubc1b\uc2b5\ub2c8\ub2e4.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\uc0ac\uc6a9\uc790 \uc804\uccb4 \uc870\ud68c",
                "description": "\ub4f1\ub85d\ub41c \ubaa8\ub4e0 \uc0ac\uc6a9\uc790 \ubaa9\ub85d\uc744 \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\uc0ac\uc6a9\uc790 \ub2e8\uac74 \uc870\ud68c",
                "description": "ID\ub85c \ud2b9\uc815 \uc0ac\uc6a9\uc790\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "\uc0ac\uc6a9\uc790 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "type": "object",
                "description": "\ub85c\uadf8\uc778 \uc694\uccad",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "\uc0ac\uc6a9\uc790 \uc774\uba54\uc77c",
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "\ube44\ubc00\ubc88\ud638",
                        "example": "password123"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "LoginResponse": {
                "type": "object",
                "description": "\ub85c\uadf8\uc778 \uc751\ub2f5",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "description": "JWT \uc561\uc138\uc2a4 \ud1a0\ud070"
                    },
                    "tokenType": {
                        "type": "string",
                        "description": "\ud1a0\ud070 \ud0c0\uc785",
                        "example": "Bearer"
                    }
                },
                "required": [
                    "accessToken",
                    "tokenType"
                ]
            },
            "UserResponse": {
                "type": "object",
                "description": "\uc0ac\uc6a9\uc790 \uc815\ubcf4",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "\uc0ac\uc6a9\uc790 ID",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "\uc0ac\uc6a9\uc790 \uc774\ub984",
                        "example": "\ud64d\uae38\ub3d9"
                    },
                    "email": {
                        "type": "string",
                        "description": "\uc0ac\uc6a9\uc790 \uc774\uba54\uc77c",
                        "example": "admin@example.com"
                    },
                    "role": {
                        "type": "string",
                        "description": "\uc5ed\ud560",
                        "example": "ADMIN"
                    }
                },
                "required": [
                    "email",
                    "id",
                    "name",
                    "role"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}